requirements.txt
Django==1.10.3
djangorestframework==3.5.3
psycopg2==2.6.2
SQLAlchemy==1.1.4